+++ src/eu/trentorise/smartcampus/eb/fragments/experience/EditExpFragment.java
@@ -329,12 +329,9 @@ public class EditExpFragment extends SherlockFragment
 
 	@Override
 	public boolean onOptionsItemSelected(MenuItem item) {
-		switch (item.getItemId()) {
-		case android.R.id.home:
+		if (item.getItemId() == android.R.id.home) {
 			getActivity().onBackPressed();
-			break;
-		case R.id.expmenu_done:
-		case R.id.expmenu_save:
+		} else if (item.getItemId() == R.id.expmenu_done || item.getItemId() == R.id.expmenu_save) {
 			exp.setTitle(mTitleSwitch.getValue());
 			exp.setDescription(mDescrSwitch.getValue());
 			if (validate(exp)) {
@@ -352,64 +349,49 @@ public class EditExpFragment extends SherlockFragment
 							"Error creating filestorage account");
 				}
 			}
-			break;
-		case R.id.expmenu_attach_audio:
+		} else if (item.getItemId() == R.id.expmenu_attach_audio) {
 			onCaptureOption(new String[] { CATCH_TYPES.AUDIO.toString() }, null);
-			break;
-		case R.id.expmenu_attach_camera_photo:
+		} else if (item.getItemId() == R.id.expmenu_attach_camera_photo) {
 			onCaptureOption(
 					new String[] { CATCH_TYPES.IMAGE_CAMERA.toString() }, null);
-			break;
-		case R.id.expmenu_attach_camera_video:
+		} else if (item.getItemId() == R.id.expmenu_attach_camera_video) {
 			onCaptureOption(
 					new String[] { CATCH_TYPES.VIDEO_CAMERA.toString() }, null);
-			break;
-		case R.id.expmenu_attach_gallery_photo:
+		} else if (item.getItemId() == R.id.expmenu_attach_gallery_photo) {
 			onCaptureOption(
 					new String[] { CATCH_TYPES.IMAGE_GALLERY.toString() }, null);
-			break;
-		case R.id.expmenu_attach_gallery_video:
+		} else if (item.getItemId() == R.id.expmenu_attach_gallery_video) {
 			onCaptureOption(
 					new String[] { CATCH_TYPES.VIDEO_GALLERY.toString() }, null);
-			break;
-		case R.id.expmenu_attach_qrcode:
+		} else if (item.getItemId() == R.id.expmenu_attach_qrcode) {
 			onCaptureOption(new String[] { CATCH_TYPES.QRCODE.toString() },
 					null);
-			break;
-		case R.id.expmenu_attach_text:
+		} else if (item.getItemId() == R.id.expmenu_attach_text) {
 			DialogFragment textFragment = new EditNoteFragment();
 			textFragment.setArguments(EditNoteFragment.prepare("", exp
 					.getContents().size()));
 			textFragment.show(getActivity().getSupportFragmentManager(),
 					"exp_content_note");
-			break;
-		case R.id.expmenu_remove:
+		} else if (item.getItemId() == R.id.expmenu_remove) {
 			DialogFragment newFragment = new DeleteExperienceFragment();
 			newFragment.setArguments(DeleteExperienceFragment.prepare(exp
 					.getId()));
 			newFragment.show(getActivity().getSupportFragmentManager(),
 					"exp_delete");
-			break;
-		case R.id.expmenu_assign_collection:
+		} else if (item.getItemId() == R.id.expmenu_assign_collection) {
 			DialogFragment assignFragment = new AssignCollectionFragment();
 			assignFragment.setArguments(AssignCollectionFragment.prepare(
 					exp.getId(), exp.getCollectionIds()));
 			assignFragment.show(getActivity().getSupportFragmentManager(),
 					"exp_assign_colls");
-			break;
-		case R.id.expmenu_share:
+		} else if (item.getItemId() == R.id.expmenu_share) {
 			ShareEntityObject obj = new ShareEntityObject(exp.getEntityId(),
 					exp.getTitle(), Constants.ENTITY_TYPE_EXPERIENCE);
 			SharingHelper.share(getActivity(), obj);
-			break;
-		case R.id.expmenu_map:
-		case R.id.expmenu_export:
+		} else if (item.getItemId() == R.id.expmenu_map || item.getItemId() == R.id.expmenu_export) {
 			Toast.makeText(getActivity(), R.string.not_implemented,
 					Toast.LENGTH_SHORT).show();
-			// TODO
-			break;
-		default:
-			break;
+		} else {
 		}
 		return super.onOptionsItemSelected(item);
 	}
@@ -655,8 +637,7 @@ public class EditExpFragment extends SherlockFragment
 				.getMenuInfo();
 		if (info.position <= exp.getContents().size() && info.position > 0) {
 			final Content content = exp.getContents().get(info.position - 1);
-			switch (item.getItemId()) {
-			case R.id.expcontentmenu_remove:
+			if (item.getItemId() == R.id.expcontentmenu_remove) {
 				switchToEdit();
 				Content c = null;
 				try {
@@ -668,8 +649,7 @@ public class EditExpFragment extends SherlockFragment
 									.getId() : null);
 				}
 				adapter.notifyDataSetChanged();
-				break;
-			case R.id.expcontentmenu_comment:
+			} else if (item.getItemId() == R.id.expcontentmenu_comment) {
 				DialogFragment newFragment = new EditNoteFragment();
 				String text = content.getType() == ContentType.TEXT ? content
 						.getValue() : content.getNote();
@@ -677,9 +657,7 @@ public class EditExpFragment extends SherlockFragment
 						info.position - 1));
 				newFragment.show(getActivity().getSupportFragmentManager(),
 						"exp_content_note");
-				break;
-			default:
-				break;
+			} else {
 			}
 		}
 		return super.onContextItemSelected(item);
